requirements.txt
astroid==2.2.5
docutils==0.14
entrypoints==0.3
flake8==3.7.7
gitdb2==2.0.5
GitPython==2.1.11
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
six==1.12.0
smmap2==2.0.5
typed-ast==1.3.1
wrapt==1.11.1
yapf==0.26.0
